requirements.txt
alembic==0.8.2
bcrypt==2.0.0
blinker==1.4
cffi==1.2.1
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.2
MarkupSafe==0.23
psycopg2==2.6.1
pycparser==2.14
pycrypto==2.6.1
PyJWT==1.4.0
python-editor==0.4
PyYAML==3.11
six==1.10.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
